write/requirements.txt
arrow==1.3.0
pillow==11.1.0
python-dateutil==2.9.0.post0
schedule==1.2.2
six==1.17.0
types-python-dateutil==2.9.0.20241206
